# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bank cards API
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearer-jwt: []
...
paths:
...
  /api/v1/admin/cards:
    get:
      tags:
        - Admin Card Management
      summary: Get all user cards
      description: Retrieve a paginated list of all cards in the system
      operationId: get
      parameters:
        - name: status
          in: query
          description: Status of the card to filter by
          required: false
          schema:
            type: string
            description: Status of the card to filter by
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
        - name: userId
          in: query
          description: Filter cards by owner user ID
          required: false
          schema:
            type: string
            format: uuid
            description: Filter cards by owner user ID
            example: d290f1ee-6c54-4b01-90e6-d701748f0851
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardDto'
    post:
      tags:
        - Admin Card Management
      summary: Create a card for a user
      description: Create a new bank card for a specified user
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
...
  /api/v1/admin/cards/{id}:
    delete:
      tags:
        - Admin Card Management
      summary: Delete a card (admin only)
      description: Soft-delete a card by its ID
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
components:
  schemas:
...
    CardDto:
      type: object
      description: Information about a user card
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the card
          example: 2e0919f1-957b-4db4-b42e-424a35dc18e2
        maskedPan:
          type: string
          description: Masked PAN of the card for security
          example: '**** **** **** 1234'
        status:
          type: string
          description: Current status of the card
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          description: Current balance of the card
          example: 1000.5
        firstNameSnapshot:
          type: string
          description: First name of the card owner (snapshot at card creation)
          example: John
        lastNameSnapshot:
          type: string
          description: Last name of the card owner
          example: Doe
        expiryMonth:
          type: integer
          format: int32
          description: Expiry month of the card
          example: 12
        expiryYear:
          type: integer
          format: int32
          description: Expiry year of the card
          example: 34
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the card was created
...
    CreateCardDto:
      type: object
      description: Request payload for creating a new card (admin only)
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user who will own the card
        pan:
          type: string
          description: Primary Account Number (PAN) of the card
          example: 4242424242424242
        firstNameSnapshot:
          type: string
          description: First name of the card owner (snapshot at creation)
          example: John
        lastNameSnapshot:
          type: string
          description: Last name of the card owner (snapshot at creation)
          example: Doe
        expiryMonth:
          type: integer
          format: int32
          description: Expiry month of the card (1-12)
          example: 12
          maximum: 12
          minimum: 1
        expiryYear:
          type: integer
          format: int32
          description: Expiry year of the card
          example: 34
      required:
        - expiryMonth
        - expiryYear
        - firstNameSnapshot
        - lastNameSnapshot
        - userId
    PageCardDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    bearer-jwt:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT